Refactor main plugin file into focused modules
   Extract birthday management, plugin integration, and view management logic from main.ts into
   separate modules for better code organization and maintainability.
